services/engine/requirements.txt
pydantic>=2.7
psycopg2-binary>=2.9
python-dotenv>=1.0
redis>=5.0
loguru>=0.7
requests>=2.31
